Day 7, Handy Haversacks, FINALLY!
Had the worst time debugging why the count for part2 was wrong. Thought I had parsed the inputs
wrong or left out part of the tree when I took the subgraph for shiny gold, but no. I forgot to
count the container bag itself, so 2 bags with 2 bags inside each became 2*2 not 2 + 2*2. :facepalm: